separate errs
