+++ src/shared/rest/resp.go
@@ -3,8 +3,6 @@ package rest
 import (
 	"encoding/json"
 	"io"
-	"strconv"
-	"time"
 
 	"github.com/spring1843/chat-server/libs/go-restful"
 )
@@ -20,24 +18,7 @@ type (
 		} `json:"metadata"`
 		Errors []RespError `json:"errors"`
 	}
-	// RespError is an error that is part of the API response
-	RespError struct {
-		Severity             int    `json:"severity"`
-		HumanFriendlyMessage string `json:"human_friendly_message"`
-		ShortMessage         string `json:"short_message"`
-	}
 )
-
-// AddError adds a REST error to response
-func (r *Resp) AddError(restError RespError) {
-	r.Errors = append(r.Errors, restError)
-}
-
-// DecorateResponse is executed before a response is generated
-func (r *Resp) DecorateResponse(request *restful.Request) {
-	r.Links.Self = "//" + request.Request.Host + request.Request.URL.String()
-	r.Metadata.Generated = strconv.FormatInt(time.Now().UnixNano(), 10)
-}
 
 // ParseRequestBody parses request  body
 func ParseRequestBody(r *restful.Request, o interface{}) error {
